Create match_operator_vs_method_and_===
Total rewrite of ===-vs-=~-vs-match.rb after reading issues on main repo. Benchmark times more consistent between runs compared to original version. Ultimately, this adds a check between the match operator with $~ along with two separate test regex to highlight differences in performance when a match is or is not found.